requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.4

# Vector Store
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# Redis
redis

# LLM Provider
openai>=1.50.0

# Token counting
tiktoken>=0.8.0

# Environment
python-dotenv>=1.0.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0
tabulate>=0.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
pydantic>=2.8.0
